requirements.txt
# OpenRouter Social Proof Dark Patterns Research Study
openai>=1.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Statistical analysis
scipy>=1.11.0
statsmodels>=0.14.0
pingouin>=0.5.3

# Configuration & data
pyyaml>=6.0
tqdm>=4.65.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
